# feat(expr): add parallel execution support for LAG/LEAD window functions